code review, code quality, bug fixes, cleanup

+++ audio_be/src/commonMain/kotlin/KlangAudioRenderer.kt
@@ -27,7 +27,7 @@ class KlangAudioRenderer(
     )
 
     fun renderBlock(
-        cursorFrame: Long,
+        cursorFrame: Int,
         out: ShortArray,
     ) {
         // 1. Reset Mix Buffers
