+++ audio_bridge/src/commonMain/kotlin/KlangPlaybackSignal.kt
@@ -30,7 +30,7 @@ sealed class KlangPlaybackSignal {
         /** Number of samples that were loaded */
         val count: Int,
         /** How long preloading took in milliseconds */
-        val durationMs: Long,
+        val durationMs: Int,
     ) : KlangPlaybackSignal()
 
     /**
@@ -77,10 +77,10 @@ sealed class KlangPlaybackSignal {
             val sourceLocations: SourceLocationChain?,
         ) {
             /** Start time in milliseconds */
-            val startTimeMs get() = (startTime * 1000).toLong()
+            val startTimeMs get() = (startTime * 1000).toInt()
 
             /** End time in milliseconds */
-            val endTimeMs get() = (endTime * 1000).toLong()
+            val endTimeMs get() = (endTime * 1000).toInt()
         }
     }
